+++ plugins/standalone/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @hot-updater/standalone
 
+## 0.35.6
+
+### Patch Changes
+
+- @hot-updater/core@0.35.6
+- @hot-updater/plugin-core@0.35.6
+
+## 0.35.5
+
+### Patch Changes
+
+- @hot-updater/core@0.35.5
+- @hot-updater/plugin-core@0.35.5
+
+## 0.35.4
+
+### Patch Changes
+
+- @hot-updater/core@0.35.4
+- @hot-updater/plugin-core@0.35.4
+
 ## 0.35.3
 
 ### Patch Changes